requirements/cpu_requirements.txt
asks
ftfy
numpy
pandas
Pillow
pycld2
random-user-agent
requests
trio
tqdm
ujson
